refactor(meeting_backend/persist): extract tests to siblings (#1266)
Partial: persist.rs 1498 -> 1076 LOC. Tests split into
tests_persist.rs (244 LOC) and tests_persist_extra.rs (~190 LOC).
Widens 4 helpers + MAX_FILENAME_LEN to pub(super). Source split deferred.
