+++ src/meeting_backend/persist.rs
@@ -14,7 +14,7 @@ use crate::meeting_facilitator::{
 use super::types::{ConversationMessage, HandoffActionItem, MeetingTranscript};
 
 /// Maximum length for a sanitized filename component.
-const MAX_FILENAME_LEN: usize = 128;
+pub(super) const MAX_FILENAME_LEN: usize = 128;
 
 /// Sanitize a string for safe use as a filesystem name.
 ///
@@ -463,7 +463,7 @@ pub fn extract_action_items(messages: &[ConversationMessage]) -> Vec<HandoffActi
 }
 
 /// Try to extract an assignee from a sentence.
-fn extract_assignee(sentence: &str) -> Option<String> {
+pub(super) fn extract_assignee(sentence: &str) -> Option<String> {
     let verbs = [" will ", " should ", " needs to ", " need to ", " must "];
     for verb in &verbs {
         if let Some(idx) = sentence.to_lowercase().find(verb) {
@@ -492,7 +492,7 @@ fn extract_assignee(sentence: &str) -> Option<String> {
 }
 
 /// Extract a deadline phrase if present.
-fn extract_deadline(lower_sentence: &str) -> Option<String> {
+pub(super) fn extract_deadline(lower_sentence: &str) -> Option<String> {
     for signal in DEADLINE_SIGNALS {
         if lower_sentence.contains(signal) {
             return Some(signal.trim().to_string());
@@ -502,7 +502,7 @@ fn extract_deadline(lower_sentence: &str) -> Option<String> {
 }
 
 /// Clean up an action item description — strip leading signal labels.
-fn clean_action_description(sentence: &str) -> String {
+pub(super) fn clean_action_description(sentence: &str) -> String {
     let mut s = sentence.trim().to_string();
     let prefixes = [
         "action item:",
@@ -528,7 +528,7 @@ fn clean_action_description(sentence: &str) -> String {
 }
 
 /// Split text into sentences (simple heuristic).
-fn split_sentences(text: &str) -> Vec<String> {
+pub(super) fn split_sentences(text: &str) -> Vec<String> {
     let mut sentences = Vec::new();
     let mut current = String::new();
     for ch in text.chars() {
@@ -1072,427 +1072,4 @@ _Tip: Timebox estimation discussions — if it takes >2 min, take it offline._",
 /// Look up a template by name. Returns `None` if not found.
 pub fn find_template(name: &str) -> Option<&'static MeetingTemplate> {
     TEMPLATES.iter().find(|t| t.name.eq_ignore_ascii_case(name))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::super::types::Role;
-    use super::*;
-    use serial_test::serial;
-
-    #[test]
-    fn sanitize_basic() {
-        assert_eq!(sanitize_filename("Sprint Planning"), "Sprint_Planning");
-    }
-
-    #[test]
-    fn sanitize_path_traversal() {
-        assert_eq!(sanitize_filename("../../etc/passwd"), "etc_passwd");
-    }
-
-    #[test]
-    fn sanitize_null_bytes() {
-        assert_eq!(sanitize_filename("test\0file"), "testfile");
-    }
-
-    #[test]
-    fn sanitize_empty() {
-        assert_eq!(sanitize_filename(""), "meeting");
-    }
-
-    #[test]
-    fn sanitize_special_chars() {
-        assert_eq!(sanitize_filename("a:b*c?d<e>f|g"), "a_b_c_d_e_f_g");
-    }
-
-    #[test]
-    fn sanitize_long_string() {
-        let long = "a".repeat(200);
-        let result = sanitize_filename(&long);
-        assert!(result.len() <= MAX_FILENAME_LEN);
-    }
-
-    #[test]
-    fn sanitize_only_dots_and_underscores() {
-        assert_eq!(sanitize_filename("...___..."), "meeting");
-    }
-
-    #[test]
-    fn find_template_by_name() {
-        assert!(find_template("standup").is_some());
-        assert!(find_template("1on1").is_some());
-        assert!(find_template("retro").is_some());
-        assert!(find_template("planning").is_some());
-        assert!(find_template("nonexistent").is_none());
-    }
-
-    #[test]
-    fn find_template_case_insensitive() {
-        assert!(find_template("STANDUP").is_some());
-        assert!(find_template("Retro").is_some());
-    }
-
-    #[test]
-    fn templates_have_content() {
-        for t in TEMPLATES {
-            assert!(!t.name.is_empty());
-            assert!(!t.description.is_empty());
-            assert!(!t.agenda.is_empty());
-        }
-    }
-
-    #[test]
-    fn at_least_four_templates() {
-        assert!(TEMPLATES.len() >= 4);
-    }
-
-    #[test]
-    fn markdown_export_format() {
-        // Verify the markdown format contains expected YAML frontmatter
-        let topic = "Test Topic";
-        let started_at = "2025-01-01T00:00:00Z";
-        let mut md = String::new();
-        md.push_str("---\n");
-        md.push_str(&format!("topic: \"{topic}\"\n"));
-        md.push_str(&format!("date: \"{started_at}\"\n"));
-        md.push_str("participants:\n  - \"operator\"\n  - \"simard\"\n");
-        md.push_str("---\n\n");
-        md.push_str(&format!("# Meeting: {topic}\n\n"));
-
-        assert!(md.contains("---"));
-        assert!(md.contains("topic: \"Test Topic\""));
-        assert!(md.contains("date: \"2025-01-01T00:00:00Z\""));
-        assert!(md.contains("participants:"));
-    }
-
-    // ── Action item extraction tests ────────────────────────────────
-
-    fn make_msg(role: Role, content: &str) -> ConversationMessage {
-        ConversationMessage {
-            role,
-            content: content.to_string(),
-            timestamp: "2025-01-01T00:00:00Z".to_string(),
-        }
-    }
-
-    #[test]
-    fn extract_action_items_from_will_verb() {
-        let messages = vec![make_msg(
-            Role::User,
-            "Alice will write the integration tests by Friday.",
-        )];
-        let items = extract_action_items(&messages);
-        assert!(!items.is_empty(), "should extract at least one action item");
-        assert_eq!(items[0].assignee.as_deref(), Some("Alice"));
-        assert_eq!(items[0].deadline.as_deref(), Some("by friday"));
-    }
-
-    #[test]
-    fn extract_action_items_labeled_prefix() {
-        let messages = vec![make_msg(
-            Role::Assistant,
-            "Action item: Deploy the staging environment.",
-        )];
-        let items = extract_action_items(&messages);
-        assert!(!items.is_empty());
-        assert!(items[0].description.contains("Deploy"));
-        assert!(!items[0].description.starts_with("Action item:"));
-    }
-
-    #[test]
-    fn extract_action_items_no_false_positives() {
-        let messages = vec![
-            make_msg(Role::User, "The weather is nice today."),
-            make_msg(Role::Assistant, "I agree, it is nice."),
-        ];
-        let items = extract_action_items(&messages);
-        assert!(items.is_empty(), "no action items in casual chat");
-    }
-
-    #[test]
-    fn extract_action_items_needs_to_pattern() {
-        let messages = vec![make_msg(
-            Role::User,
-            "Bob needs to update the CI pipeline this week.",
-        )];
-        let items = extract_action_items(&messages);
-        assert!(!items.is_empty());
-        assert_eq!(items[0].assignee.as_deref(), Some("Bob"));
-        assert_eq!(items[0].deadline.as_deref(), Some("this week"));
-    }
-
-    #[test]
-    fn extract_assignee_from_assigned_to() {
-        let result = extract_assignee("This task is assigned to Carol for next sprint.");
-        assert_eq!(result.as_deref(), Some("Carol"));
-    }
-
-    #[test]
-    fn extract_deadline_various() {
-        assert_eq!(extract_deadline("do it by eod"), Some("by eod".to_string()));
-        assert_eq!(
-            extract_deadline("finish next sprint"),
-            Some("next sprint".to_string())
-        );
-        assert_eq!(extract_deadline("nothing here"), None);
-    }
-
-    #[test]
-    fn clean_action_description_strips_prefixes() {
-        assert_eq!(
-            clean_action_description("TODO: Fix the tests"),
-            "Fix the tests"
-        );
-        assert_eq!(clean_action_description("task: Review PR"), "Review PR");
-        assert_eq!(clean_action_description("Normal text"), "Normal text");
-    }
-
-    #[test]
-    fn split_sentences_basic() {
-        let sentences = split_sentences("Hello world. How are you? Fine!");
-        assert_eq!(sentences.len(), 3);
-    }
-
-    // ── Goal linkage tests ──────────────────────────────────────────
-
-    #[test]
-    fn link_action_items_exact_overlap() {
-        let mut items = vec![HandoffActionItem {
-            description: "Set up continuous integration pipeline".to_string(),
-            assignee: None,
-            deadline: None,
-            linked_goal: None,
-            priority: None,
-        }];
-        let goals = vec![(
-            "ci-pipeline".to_string(),
-            "Set up continuous integration".to_string(),
-        )];
-        link_action_items_to_goals(&mut items, &goals);
-        assert_eq!(items[0].linked_goal.as_deref(), Some("ci-pipeline"));
-    }
-
-    #[test]
-    fn link_action_items_no_match() {
-        let mut items = vec![HandoffActionItem {
-            description: "Order new keyboards".to_string(),
-            assignee: None,
-            deadline: None,
-            linked_goal: None,
-            priority: None,
-        }];
-        let goals = vec![(
-            "improve-testing".to_string(),
-            "Improve testing coverage".to_string(),
-        )];
-        link_action_items_to_goals(&mut items, &goals);
-        assert!(items[0].linked_goal.is_none());
-    }
-
-    // ── Decision extraction tests ───────────────────────────────────
-
-    #[test]
-    fn extract_decisions_from_transcript() {
-        let messages = vec![
-            make_msg(Role::User, "I think we should use Rust."),
-            make_msg(
-                Role::Assistant,
-                "Decision: We will adopt Rust for the backend.",
-            ),
-            make_msg(Role::User, "We agreed to ship by end of month."),
-        ];
-        let decisions = extract_decisions(&messages);
-        assert!(decisions.len() >= 2, "got: {decisions:?}");
-        assert!(decisions.iter().any(|d| d.contains("Rust")));
-        assert!(decisions.iter().any(|d| d.contains("agreed")));
-    }
-
-    #[test]
-    fn extract_decisions_none_found() {
-        let messages = vec![
-            make_msg(Role::User, "Let's discuss options."),
-            make_msg(Role::Assistant, "Here are some possibilities."),
-        ];
-        let decisions = extract_decisions(&messages);
-        assert!(decisions.is_empty());
-    }
-
-    // ── Open question extraction tests ──────────────────────────────
-
-    #[test]
-    fn extract_open_questions_explicit_markers() {
-        let messages = vec![
-            make_msg(Role::User, "OPEN: What database should we use?"),
-            make_msg(Role::Assistant, "Question: Who owns the rollback plan?"),
-        ];
-        let questions = extract_open_questions(&messages);
-        assert_eq!(questions.len(), 2);
-        assert!(questions[0].explicit);
-        assert!(questions[1].explicit);
-    }
-
-    #[test]
-    fn extract_open_questions_genuine_question() {
-        let messages = vec![make_msg(
-            Role::User,
-            "How should we handle backward compatibility for the API?",
-        )];
-        let questions = extract_open_questions(&messages);
-        assert_eq!(questions.len(), 1);
-        assert!(!questions[0].explicit);
-        assert!(questions[0].text.contains("backward compatibility"));
-    }
-
-    #[test]
-    fn extract_open_questions_skips_short() {
-        let messages = vec![make_msg(Role::User, "Why not?")];
-        let questions = extract_open_questions(&messages);
-        assert!(
-            questions.is_empty(),
-            "Short rhetorical-like questions should be filtered"
-        );
-    }
-
-    #[test]
-    fn extract_open_questions_empty_messages() {
-        let questions = extract_open_questions(&[]);
-        assert!(questions.is_empty());
-    }
-
-    // ── Theme extraction tests ──────────────────────────────────────
-
-    #[test]
-    fn extract_themes_recurring_words() {
-        let messages = vec![
-            make_msg(Role::User, "We need to improve testing coverage."),
-            make_msg(Role::Assistant, "Testing is important for quality."),
-            make_msg(Role::User, "Let's add more testing to the pipeline."),
-        ];
-        let themes = extract_themes(&messages);
-        assert!(
-            themes.contains(&"testing".to_string()),
-            "Expected 'testing' in themes: {themes:?}"
-        );
-    }
-
-    #[test]
-    fn extract_themes_empty_messages() {
-        let themes = extract_themes(&[]);
-        assert!(themes.is_empty());
-    }
-
-    #[test]
-    fn extract_themes_skips_system_messages() {
-        let messages = vec![
-            make_msg(
-                Role::System,
-                "System prompt with repeated system words system.",
-            ),
-            make_msg(Role::User, "Hello"),
-        ];
-        let themes = extract_themes(&messages);
-        // "system" only appeared in system messages, which are skipped.
-        assert!(!themes.contains(&"system".to_string()));
-    }
-
-    // ── write_handoff completeness test ─────────────────────────────
-
-    #[test]
-    #[serial]
-    fn write_handoff_includes_structured_data() {
-        let dir = tempfile::tempdir().unwrap();
-        unsafe {
-            std::env::set_var("SIMARD_HANDOFF_DIR", dir.path().as_os_str());
-        }
-
-        let messages = vec![
-            make_msg(Role::User, "We need better testing."),
-            make_msg(
-                Role::Assistant,
-                "Decision: We will adopt TDD. OPEN: Who will lead the effort?",
-            ),
-        ];
-        let action_items = vec![HandoffActionItem {
-            description: "Set up CI pipeline".to_string(),
-            assignee: Some("alice".to_string()),
-            deadline: Some("Friday".to_string()),
-            linked_goal: None,
-            priority: None,
-        }];
-        let decisions = vec!["We will adopt TDD".to_string()];
-
-        let result = write_handoff(
-            "Sprint planning",
-            "Good meeting",
-            &messages,
-            &action_items,
-            &decisions,
-        );
-        assert!(result.is_ok(), "write_handoff failed: {result:?}");
-
-        // Read the written handoff file.
-        let entries: Vec<_> = std::fs::read_dir(dir.path())
-            .unwrap()
-            .filter_map(|e| e.ok())
-            .collect();
-        assert!(!entries.is_empty(), "No handoff file written");
-
-        let content = std::fs::read_to_string(entries[0].path()).unwrap();
-        let handoff: MeetingHandoff = serde_json::from_str(&content).unwrap();
-
-        // Decisions are populated.
-        assert_eq!(handoff.decisions.len(), 1);
-        assert!(handoff.decisions[0].description.contains("TDD"));
-
-        // Action items are populated.
-        assert_eq!(handoff.action_items.len(), 1);
-        assert_eq!(handoff.action_items[0].description, "Set up CI pipeline");
-        assert_eq!(handoff.action_items[0].owner, "alice");
-
-        // Open questions are extracted from messages.
-        assert!(
-            !handoff.open_questions.is_empty(),
-            "Expected open questions from message content"
-        );
-
-        // Participants include roles from messages and assignees.
-        assert!(handoff.participants.contains(&"operator".to_string()));
-        assert!(handoff.participants.contains(&"alice".to_string()));
-
-        // Transcript contains summary.
-        assert!(handoff.transcript.contains(&"Good meeting".to_string()));
-
-        unsafe {
-            std::env::remove_var("SIMARD_HANDOFF_DIR");
-        }
-    }
-
-    #[test]
-    #[serial]
-    fn write_handoff_empty_data_uses_defaults() {
-        let dir = tempfile::tempdir().unwrap();
-        unsafe {
-            std::env::set_var("SIMARD_HANDOFF_DIR", dir.path().as_os_str());
-        }
-
-        let result = write_handoff("Empty meeting", "No notes", &[], &[], &[]);
-        assert!(result.is_ok());
-
-        let entries: Vec<_> = std::fs::read_dir(dir.path())
-            .unwrap()
-            .filter_map(|e| e.ok())
-            .collect();
-        let content = std::fs::read_to_string(entries[0].path()).unwrap();
-        let handoff: MeetingHandoff = serde_json::from_str(&content).unwrap();
-
-        assert!(handoff.decisions.is_empty());
-        assert!(handoff.action_items.is_empty());
-        assert!(handoff.open_questions.is_empty());
-        assert!(handoff.participants.is_empty());
-        assert!(handoff.themes.is_empty());
-
-        unsafe {
-            std::env::remove_var("SIMARD_HANDOFF_DIR");
-        }
-    }
 }
